requirements.txt
torch==2.1.0
Pillow
diffusers
transformers
accelerate
scipy
safetensors
imageio